renamed Minecraft Files to Instance files
