+++ src/components/Instance/index.ts
@@ -1,0 +1,4 @@
+export { default as InstancePerformanceCard } from './InstancePerformanceCard';
+export { default as InstanceGeneralCard } from './InstanceGeneralCard';
+export { default as InstanceSettingCard } from './InstanceSettingCard';
+export { default as InstancePlayerList } from './InstancePlayerList';
