--- src/components/Minecraft/index.ts
@@ -1,4 +1,0 @@
-export { default as MinecraftPerformanceCard } from './MinecraftPerformanceCard';
-export { default as MinecraftGeneralCard } from './MinecraftGeneralCard';
-export { default as MinecraftSettingCard } from './MinecraftSettingCard';
-export { default as MinecraftPlayerList } from './MinecraftPlayerList';
